Drone Radar
You have to run main.py first to setup the config file and then run drone.py to run the webserver. This will then scout out players.

If there are any issues, open an issue on GitHub and I'll take a look (trying to currently work on accuracy, seems to be only the tiniest bit off tho, NPCs shouldn't be picked up).